refactor(components): rename files and update references
Renamed component and style files to follow PascalCase naming convention for consistency. Updated all relevant imports and references to reflect the changes. Fixed typos in naming (e.g., "caroussel" to "carousel") for standardization.
