+++ src/components/Carousel.tsx
