+++ src/components/Divided.tsx
