+++ src/components/ProgressIndicator.tsx
