+++ src/components/Slider.tsx
